requirements.txt
Django==2.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytube3==9.6.4
pytz==2019.3
sqlparse==0.3.1
typing-extensions==3.7.4.2